Fix Divide by Zero bug in translateToPx()
Bug that drove me insane hunting down and fixing back in 2010.  Found
the bug is still in the latest version so I applied it to the latest
branch.  If I recall correctly it fails SILENTLY with no errors and
sliders that just don't work. Took days to initially find the problem.